waiting for locksmith and paywall to be up before running  the test suite
