+++ tests/README.md
@@ -32,6 +32,7 @@ what is being rendered on screen.
 - start a local ganache node (at the root): `npm run start-ganache -- -b 1` (the -b 1 means that the node will mine blocks every second, even if no transaction has been added)
 - deploy the smart contracts (in `/unlock-app`): `npm run deploy-unlock-contract`
 - run the dashboard application (in `/unlock-app`): `npm run start` (you may need to build the application first: `npm run build`)
+- run the locksmith application (in `/locksmith`): `npm run start`
 - run the paywall application (in `/paywall`): `npm run start` (you may need to build the application first: `npm run build`)
 - execute the tests (in `/tests`): `npm run test`
